+++ ui/PinYin.py
@@ -3,10 +3,11 @@ import os
 import subprocess
 from sys import platform
 
-import fitz
+from pdf2image import convert_from_path
 from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QTextOption, QFontDatabase, QFont, QImage, QPixmap
-from PyQt5.QtWidgets import QWidget, QGridLayout, QLabel, QLineEdit, QComboBox, QTextEdit, QPushButton, QHBoxLayout
+from PyQt5.QtWidgets import QWidget, QGridLayout, QLabel, QLineEdit, QComboBox, QTextEdit, QPushButton, QHBoxLayout, \
+    QSizePolicy
 
 from backend.PinYin import PinYin
 
@@ -103,9 +104,12 @@ class UiPinYin(QWidget):
         self.btn_ok = QPushButton("  生成拼音字帖  ")
         self.btn_ok.clicked.connect(self.on_click_ok)
         grid.addWidget(self.btn_ok, row, 1, alignment=Qt.AlignCenter)
-
+        self.container = QWidget(self)
+        self.container.setLayout(grid)
+        self.container.setMinimumWidth(640)
+        self.container.setSizePolicy(QSizePolicy.Fixed, QSizePolicy.Expanding)
         hbox = QHBoxLayout()
-        hbox.addLayout(grid)
+        hbox.addWidget(self.container)
 
         self.preview = QLabel()
         hbox.addWidget(self.preview)
@@ -117,6 +121,10 @@ class UiPinYin(QWidget):
         self.combo_types.currentIndexChanged.connect(self.do_preview)
         self.edit_text.textChanged.connect(self.do_preview)
         self.do_preview()
+
+    def resizeEvent(self, a0):
+        self.do_preview()
+        super().resizeEvent(a0)
 
     def on_click_ok(self):
         self.btn_ok.setEnabled(False)
@@ -156,14 +164,20 @@ class UiPinYin(QWidget):
     def do_preview(self):
         pdf_path = "/tmp/1.pdf"
         self.do_draw(pdf_path)
-        with fitz.open(pdf_path) as doc:
-            if doc.page_count <= 0:
-                return
-            page = doc[0]
-            img = page.get_pixmap()
+        images = convert_from_path(pdf_path, fmt='png', dpi=72, last_page=1)
+        if len(images) <= 0:
+            return
+        img = images[0]
 
         os.remove(pdf_path)
-        fmt = QImage.Format_RGBA8888 if img.alpha else QImage.Format_RGB888
-        img = QImage(img.samples, img.width, img.height, img.stride, fmt)
-        self.preview.setPixmap(QPixmap(img))
-        self.preview.setFixedWidth(img.width())
+        img = img.convert("RGBA")
+        data = img.tobytes("raw", "BGRA")
+        im = QImage(data, img.width, img.height, QImage.Format.Format_ARGB32)
+
+        h = int(self.preview.width() / img.width * img.height)
+        if h > self.preview.height():
+            im1 = im.scaledToHeight(self.preview.height(), Qt.SmoothTransformation)
+        else:
+            im1 = im.scaledToWidth(self.preview.width(), Qt.SmoothTransformation)
+
+        self.preview.setPixmap(QPixmap(im1))
